Added a few buttons to the Activities in order to be able to navigate between them.
I will work on the flow of the App briefly a bit later, but the idea is to be able to click on an App and be navigated to the details regarding that App. I am hoping these details will include things like IP address and port numbers for destination and source.