Defining our Models
In order to access our records from the database more easily, we should create a model class for each of our resources. In this case, let's define a Post and a User model:

These are just testing classes and will later be replaced with the correct Classes for our Machine Learning Malware Detection App.