Commiting changes, after commenting out some of my code to allow the project to compile.
I am goign to hold off on the VPN Service and only focus on parsing the default statistics provided by the Android Platform, so that Weka can use it in the Machine Learning. That way, I can get results to include on my paper and provide Professor Geetha. I will then come back and add the rest later.